Log statement bug fixed, added activity bar to when scanning
When withdrawing, the log statement still said deposit, now fixed to saying withdraw.

When scanning the card, if the card's ID was not in the CSV file then it will do nothing, now the activity bar will state bad read or invalid card. Furthermore if the COM port could not open it would also do nothing but now the activity bar will state "Failed to open port" + COMN port.